third_party/blink/web_tests/accessibility/aria-checkbox-checked-expected.txt

 
X
X
   

This tests that ARIA checkboxes correctly handle the aria-checked attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS checkbox.checked is "false"
PASS checkbox.checked is "true"
PASS checkbox.checked is "false"
PASS checkbox.checked is "true"
PASS checkbox.checked is "false"
PASS checkbox.checked is "true"
PASS checkbox.checked is "false"
PASS checkbox.checked is "true"
PASS successfullyParsed is true

TEST COMPLETE
